Move simple java mail sender logics to EmailUtill class
