+++ src/main/java/org/sefglobal/scholarx/service/EmailService.java
@@ -1,6 +1,7 @@
 package org.sefglobal.scholarx.service;
 
 import org.sefglobal.scholarx.model.Mail;
+import org.sefglobal.scholarx.util.EmailUtil;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
@@ -10,21 +11,21 @@ import java.util.Map;
 @Service
 public class EmailService {
 
-    private SimpleJavaMail simpleJavaMail;
+    private EmailUtil emailUtil;
 
     public void sendEmail(String emailAddress, String subject, String message, boolean showButton) throws IOException {
-        Mail email = new Mail();
-        email.setEmail(emailAddress);
-        email.setSubject(subject);
-        email.setMessage(message);
+        Mail mail = new Mail();
+        mail.setEmail(emailAddress);
+        mail.setSubject(subject);
+        mail.setMessage(message);
 
         Map<String, Object> model = new HashMap<String, Object>();
         model.put("emailAddress", emailAddress);
         model.put("subject", subject);
         model.put("message", message);
         model.put("showButton", showButton);
-        email.setProps(model);
+        mail.setProps(model);
 
-        simpleJavaMail.sendEmail(email);
+        emailUtil.sendEmail(mail);
     }
 }
